BetterPawnControl Compatible assembly version
I recompiled this one from the latest version of Voult's BPC with additional modifications so as to integrate it within the mod. Replace this assembly into the one  you currently have in BPC mod folder and your policies will also affect Weapon Presets . Remember to make a backup of the original assembly..
It works but needs testing anyway.